WIP: configure/deploy via helm charts
This commit is an example of using charts to deploy each
openwhisk component via charts instead of managing with ansible.

Taken a little bit further the entire application could be
deployable using a single chart (which references other charts).
All of the logic in the ansible scripts can be constructed with
kubernetes and helm functionality.

For instance deploying nginx is now:

	helm install --name demo --set nginx.domainName=localhost charts/nginx

The certs are generated by an init container automatically.

This is a WIP.  I wanted to open it up for review/comments.