enable active-active controller replicas (#85)
Switch to active-active controller replication for Kubernetes. This
enables controller scale out and also simplifies the deployment
configuration be reducing the number of places that need to hardwire
the exact number of controllers.

With this change in place, I can kill a controller and have
the system (a) not hang trying to talk to the dead controller
and (b) start using the replacement controller instance once
it is available.