Update READMEs and kafka image to exploit PR-2948 (#89)
The upstream PR-2948 enabled lazy creation of the
invokerN and completedN topics, so they no longer
needed to be created by the init.sh script in the
kafka image.  In conjunction with the change in
PR-85 to switch to active-active controllers, this
means that changing the number of deployed invokers
or controllers is localized to just the invoker.yml
or controller.yml file.

Also switch custom kafka image to dgrove/whisk_kafka
for easier updates and version bump parent image
to ches/kafka 0.10.2.1 to match openwhisk upstream.