support routemgmt installation for apigateway
Add support for running a Job to deploy the routemgmt functions
in support of apigateway functionality.  The functionality is not
mentioned in the toplevel README.md or in the travis build script
yet because we first need a successful travis build on master to
publish the new openwhisk/kube-routemgmt image to dockerhub.