bugfix for setting WSK_HOST on minikube
When minikube isn't using the none driver, the Hostname
returned from `kubectl describe nodes` may not be resolvable.
Use `minikube ip` instead to get an ip addr for WSK_HOST.

Also update the logs command that runs when the script fails
to use the name=invoker label (missed when we converted the
invoker from statefulsets to daemonsets).