fix Travis CI regression introduced by 84f6e38d.
Only use `minikube ip` to get WSK_HOST when `kubectl describe nodes`
returns "minikube" as the hostname.  When running with the none driver
(like in travis CI), minikube ip returns local host, which works
for the wsk CLI externally, but not for the wsk CLI when invoked
from a Job running within minikube.

Regression was not detected during PR testing because the fix
of 63f96636 to check error conditions in the job's init.sh has not
been merged to master when 84f6e38d was tested.