configurable timeout for travis/build.sh
Add hook to allow overriding the timeout value used when waiting for a
pod to be successfully deployed.  I've found when running locally on a
fresh minikube VM, it can sometimes take an excessively long time to
pull images from dockerhub into the docker registry within the
minikube cluster.  Setting a high timeout limit avoids needing to run
the script multiple times to pull in the needed images.