Helm support for clustered zookeeper
Allow zookeeper.replicaCount to be set to values > 1 by
properly configuring the zookeeper cluster and adjusting other
parts of the Helm chart so that dependent services can
properly communicate with the cluster.

Also document which of the deployments do not yet support
multiple replicas in docs/configurationChoices.md and
via comments in values.yaml.