store key OpenWhisk configuration values in ConfigMaps
Go back to using ConfigMaps to store key cluster and database
configuration parameters and modify deployments to access the data
from those maps. Having explicit config maps makes it easier to
inspect the cluster configuration and enable separating the core
system and the providers into separate charts (subsequent PR after
this is merged).

Note: database and system authentication information continues to
be stored in secrets; that is unchanged by this PR.