Move db initialization into separate Job and reorder invoker init-containers (#261)
Restructure the db initialization into a separate Job using the
whisk-ansible-runner docker image that can be optionally run on both
"in cluster" and "external" database instances. In addition to
simplifying usage of an external database, this also allows the use of
a vanilla apache/couchdb image for in cluster CouchDB.

Also reorder invoker init-containers so we do the docker-pull-runtimes
before waiting for the controller to be up. This better overlaps the
docker pulls of the runtime containers with the inherent latency of
waiting for the controller to be ready.