bug fix in wait-for-couchdb; improve logs of wait-for-X initContainers
Check for the presence of a synthetic marker database that is only created
after the ansible init_db and wipe_db jobs complete to avoid deciding
that CouchDB is ready while the wipe_db job is still actually running.

More descriptive messages from wait-for-X jobs while running
and always print a Success message when the job completes.