Enable persistence by default + reorg of install instructions
Consolidate options around persistence and enable it by default.
Kubernetes on Docker for Mac has good support for dynamic provisioning of
persistent volumes (via the hostpath provisioner), so enabling this by
default is the less surprising option.  In particular, it allows an
OpenWhisk deployment to survive controlled host system reboots and still
be usable.

Major reorganization of setup/install instructions to group by the
different kinds of Kubernetes clusters.  I think this should result in
a simpler install experience, since the Kubernetes specific details are
better organized and distinct from the main flow which is common to all
clusters.