Helm best practice: include .Release.Name in all resource names
To allow multiple instances of the Chart to be deployed on the same
cluster, it is a best practice to include .Release.Name as part of the
name of all resources defined by the Chart. It now works to deploy
multiple instances of the Chart on the same cluster (even in the same
namespace).

Remove the partial support for using external zookeeper and/or kafka
instances for the OpenWhisk control plane (was oddly using name value
as the external hostname of the service).

Small improvements in the travis collect-logs script as well.