Enforce unique keys for whisk.system and guest namespaces (#419)
Fixes #418

We require unique auth values for each subject. Check that the
whisk.auth.guest and whisk.auth.system keys are unique during
couchdb initialization and fail with a clear error message instead
of allowing an unusable system to be deployed.

Also add an explicit `rm` of the default keys right after the git clone
of the openwhisk repository to make sure they can't be used by mistake
in some subsequent script operation.