Document deploying OpenWhisk on kind
The kind project is the replacement for the retired Kubernetes
dind-cluster project. Kind uses Docker-in-Docker virtualization to
create a virtual multi-node Kubernetes cluster on a single host.

This commit contains multiple pieces:
1. Documents usage of kind to deploy OpenWhisk
2. Switches TravisCI from using dind-cluster to kind
3. Removes minikube from top-level README in favor
   of steering people to kind.

Switching to kind for TravisCI is required to be able to track
Kubernetes versions (dind-cluster stops at 1.15).

One unfortunate aspect of this switch is that kind uses containerd as
the underlying container engine, therefore it is no longer possible to
test the DockerContainerFactory in the deploy-kube TravisCI runs. This
is a symptom of the broader Kubernetes ecosystem moving away from
Docker in favor of containerd.